+++ render/player.cpp
@@ -134,15 +134,15 @@ void player::noteOff(note * n){
     }
 }
 
-int player::getTime(){
+long player::getTime(){
     auto n = std::chrono::system_clock::now();
     auto d = std::chrono::duration_cast<std::chrono::milliseconds>(n.time_since_epoch());
     return d.count();
 }
 
 void player::goNextStep(){
-    int tt    = getTime();
-    int delta = tt-lastTime;
+    long tt    = getTime();
+    long delta = tt-lastTime;
     lastTime    = tt;
     float dtick = ticksPerSecond()*(delta/1000.0);
     lookAtX+=dtick;
